Very simple single view apps for a Swift tutorial
This was created for the Texas A&M Computing Society, of which I was
formerly webmaster and am currently secretary. We host industrial and
academic technical talks, and occasionally also allow student
presentations. This week (3/22) I will be giving a brief tutorial on
Swift using one (or both!) of these applications: a color picker using
UISliders, and a simple shift cipher. All of these are simple,
single-view apps designed to familiarize the audience with Swift 2.